Updates from subscale flight

+++ lib/MPU6050/MPU6050.cpp
@@ -354,9 +354,9 @@ Vector MPU6050::readRawAccel(void)
     uint8_t zla = Wire.receive();
 #endif
 
-    ra.XAxis = xha << 8 | xla;
-    ra.YAxis = yha << 8 | yla;
-    ra.ZAxis = zha << 8 | zla;
+    ra.XAxis = (float)((int16_t)(xha << 8 | xla));
+    ra.YAxis = (float)((int16_t)(yha << 8 | yla));
+    ra.ZAxis = (float)((int16_t)(zha << 8 | zla));
 
     return ra;
 }
@@ -415,9 +415,9 @@ Vector MPU6050::readRawGyro(void)
     uint8_t zla = Wire.receive();
 #endif
 
-    rg.XAxis = xha << 8 | xla;
-    rg.YAxis = yha << 8 | yla;
-    rg.ZAxis = zha << 8 | zla;
+    rg.XAxis = (float)((int16_t)(xha << 8 | xla));
+    rg.YAxis = (float)((int16_t)(yha << 8 | yla));
+    rg.ZAxis = (float)((int16_t)(zha << 8 | zla));
 
     return rg;
 }
